refactor : s3 로직 제거, discord 메세지 Orchestrator에서 전송
