--- src/main/java/com/example/ajouevent_be_v2/config/S3Config.java
@@ -1,29 +1,0 @@
-package com.example.ajouevent_be_v2.config;
-
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.example.ajouevent_be_v2.config.properties.S3Properties;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class S3Config {
-
-    private final S3Properties s3Properties;
-
-    @Bean
-    public AmazonS3 amazonS3() {
-        BasicAWSCredentials basicAWSCredentials =
-                new BasicAWSCredentials(
-                        s3Properties.getCredentials().getAccessKey(),
-                        s3Properties.getCredentials().getSecretKey());
-        return AmazonS3ClientBuilder.standard()
-                .withRegion(s3Properties.getRegion())
-                .withCredentials(new AWSStaticCredentialsProvider(basicAWSCredentials))
-                .build();
-    }
-}
