--- src/main/java/com/example/ajouevent_be_v2/config/properties/S3Properties.java
@@ -1,30 +1,0 @@
-package com.example.ajouevent_be_v2.config.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@Component
-@ConfigurationProperties(prefix = "cloud.aws")
-public class S3Properties {
-
-    private Credentials credentials = new Credentials();
-    private String region;
-    private S3 s3 = new S3();
-
-    @Getter
-    @Setter
-    public static class Credentials {
-        private String accessKey;
-        private String secretKey;
-    }
-
-    @Getter
-    @Setter
-    public static class S3 {
-        private String bucket;
-    }
-}
